dashboard/app: truncate console output log
For fs related crashes, the console output log is trimmed if exceeds 1MB [1]
losing kernel logs at the end. Truncating strategy wit prefix and postfix of
output log should reserve this significant data.

This fix doesn't solve gzip compression problem addressed in issue [2].

[1] https://lore.kernel.org/all/CANp29Y5edqzjYyXEgeU8i9mrrfOkPBMWYd8yMnhrozePFuWhkA@mail.gmail.com/
[2] https://github.com/google/syzkaller/issues/5250
