+++ pkg/report/report.go
@@ -768,27 +768,6 @@ func replace(where []byte, start, end int, what []byte) []byte {
 	return where
 }
 
-// Truncate leaves up to `begin` bytes at the beginning of log and
-// up to `end` bytes at the end of the log.
-func Truncate(log []byte, begin, end int) []byte {
-	if begin+end >= len(log) {
-		return log
-	}
-	var b bytes.Buffer
-	b.Write(log[:begin])
-	if begin > 0 {
-		b.WriteString("\n\n")
-	}
-	fmt.Fprintf(&b, "<<cut %d bytes out>>",
-		len(log)-begin-end,
-	)
-	if end > 0 {
-		b.WriteString("\n\n")
-	}
-	b.Write(log[len(log)-end:])
-	return b.Bytes()
-}
-
 var (
 	filenameRe    = regexp.MustCompile(`([a-zA-Z0-9_\-\./]*[a-zA-Z0-9_\-]+\.(c|h)):[0-9]+`)
 	reportFrameRe = regexp.MustCompile(`.* in ([a-zA-Z0-9_]+)`)
