+++ syz-manager/manager.go
@@ -761,7 +761,7 @@ func (mgr *Manager) NeedRepro(crash *manager.Crash) bool {
 func truncateReproLog(log []byte) []byte {
 	// Repro logs can get quite large and we have trouble sending large API requests (see #4495).
 	// Let's truncate the log to a 512KB prefix and 512KB suffix.
-	return report.Truncate(log, 512000, 512000)
+	return osutil.Truncate(log, 512000, 512000)
 }
 
 func (mgr *Manager) saveFailedRepro(rep *report.Report, stats *repro.Stats) {
